Fix comment and update README.md
