+++ web/web.go
@@ -1,7 +1,12 @@
 package web
 
 import (
+	"encoding/json"
+	"fmt"
+	"github.com/polisgo2020/search-senyast4745/index"
+	"github.com/polisgo2020/search-senyast4745/util"
 	"net/http"
+	"strings"
 	"time"
 
 	"github.com/go-chi/chi"
@@ -14,20 +19,31 @@ import (
 type App struct {
 	Mux          *chi.Mux
 	netInterface string
+	Index        *index.Index
 }
 
-func NewApp(c *config.Config, middlewares ...func(handler http.Handler) http.Handler) (*App, error) {
+type FileResponse struct {
+	Filename string
+	Count    int
+	Spacing  int
+}
+
+func NewApp(c *config.Config, index *index.Index) (*App, error) {
 	r := chi.NewMux()
-	for i := range middlewares {
-		r.Use(middlewares[i])
-	}
+
+	log.Debug().Msg("add custom log middleware")
+	r.Use(logMiddleware)
+
 	d, err := time.ParseDuration(c.TimeOut)
 	if err != nil {
+		log.Warn().Str("timeout", c.TimeOut).Msg("can not parse timeout")
 		d = 10
 	}
 
+	log.Debug().Dur("timeout", d).Msg("server timeout")
+
 	r.Use(middleware.Timeout(d * time.Millisecond))
-	//filesDir := http.Dir("static")
+
 	corsFilter := cors.New(cors.Options{
 		AllowedOrigins:   []string{"*"},
 		AllowedMethods:   []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
@@ -36,33 +52,67 @@ func NewApp(c *config.Config, middlewares ...func(handler http.Handler) http.Han
 		AllowCredentials: true,
 		MaxAge:           300,
 	})
+
+	log.Debug().Interface("cors policy", corsFilter).Msg("cors policy created")
+
 	r.Use(corsFilter.Handler)
-	//fileServer(r, "/", filesDir)
-	if c.Listen == "" {
-		c.Listen = ":8888"
-	}
-	return &App{Mux: r, netInterface: c.Listen}, nil
+
+	log.Debug().RawJSON("endpoint", []byte("{\"method\" : \"POST\", \"pattern\" : \"\\\"")).Msg("register controller")
+
+	r.Post("/", func(w http.ResponseWriter, req *http.Request) {
+		searchWords := req.FormValue("search")
+		log.Info().Str("search phrase", searchWords).Msg("start search")
+		var inputWords []string
+		for _, word := range strings.Split(searchWords, " ") {
+			util.CleanUserInput(word, func(input string) {
+				inputWords = append(inputWords, input)
+			})
+		}
+		log.Debug().Msgf("clean input: %+v", inputWords)
+		if len(inputWords) == 0 {
+			log.Err(nil).Str("input", searchWords).Msg("Incorrect search words")
+			http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+			return
+		}
+
+		var resp []FileResponse
+		for k, v := range index.Search(inputWords) {
+			resp = append(resp, FileResponse{
+				Filename: k,
+				Count:    v.Path,
+				Spacing:  v.Weight,
+			})
+		}
+		log.Info().Interface("result", resp).Msgf("search finished")
+		log.Debug().Msg("start marshalling and writing data to response")
+		rawData, err := json.Marshal(resp)
+		if err != nil {
+			log.Err(err).Interface("json data", resp).Msg("error while marshalling data to json")
+			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		}
+		if _, err = fmt.Fprint(w, string(rawData)); err != nil {
+			log.Printf("error %s while writing data %s do json\n", err, string(rawData))
+			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		}
+	})
+
+	return &App{Mux: r, netInterface: c.Listen, Index: index}, nil
 }
 
-/*func fileServer(r chi.Router, path string, root http.FileSystem) {
-	if strings.ContainsAny(path, "{}*") {
-		panic("FileServer does not permit any URL parameters.")
-	}
+func logMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		start := time.Now()
+		next.ServeHTTP(w, r)
 
-	if path != "/" && path[len(path)-1] != '/' {
-		r.Get(path, http.RedirectHandler(path+"/", 301).ServeHTTP)
-		path += "/"
-	}
-	path += "*"
-
-	r.Get(path, func(w http.ResponseWriter, r *http.Request) {
-		log.Info().Msg("getting index.html on /")
-		rctx := chi.RouteContext(r.Context())
-		pathPrefix := strings.TrimSuffix(rctx.RoutePattern(), "/*")
-		fs := http.StripPrefix(pathPrefix, http.FileServer(root))
-		fs.ServeHTTP(w, r)
+		log.Debug().
+			Str("method", r.Method).
+			Str("remote", r.RemoteAddr).
+			Str("path", r.URL.Path).
+			Int("duration", int(time.Since(start))).
+			Msgf("Called url %s", r.URL.Path)
 	})
-}*/
+
+}
 
 func (a *App) Run() {
 	log.Info().Str("network interface", a.netInterface).Msg("server start")
